TurbulenceConvection.jl

TurbulenceConvection.ClimaParams — Module

ClimaParams

This is a module for staging experimental clima parameters– that we are unsure will be used in the final model. Once they are no longer experimental, they should be moved to CLIMAParameters.jl.

TODO: move non-experimental parameters to CLIMAParameters

TurbulenceConvection.ccut_downwind — Function

ccut_downwind

Used when - traversing cell centers - grabbing one-sided (downwind) stencil of cell center k and cell center k+1

This is needed for "upwinding" rain, which travels down (hence the direction change).

TurbulenceConvection.ccut_upwind — Function

ccut_upwind

Used when - traversing cell centers - grabbing one-sided (upwind) stencil of cell center k and cell center k-1

TurbulenceConvection.daul_f2c_upwind — Function

daul_f2c_upwind

Used when - traversing cell centers - grabbing interpolated one-sided (upwind) stencil of cell center k and cell center k-1

TurbulenceConvection.dual_centers — Function

dual_centers

Used when - traversing cell faces - grabbing stencil of 2 neighboring cell centers